+++ BattleCity_TeamProject_Team2/GranadeItem.cpp
@@ -9,14 +9,14 @@ HRESULT GranadeItem::Init()
 {
 	m_img = ImageManager::GetSingleton()->FindImage("Image/BattleCity/Item/Item5.bmp");
 
-	m_pos.x = TILE_MAP_START_POS_X + rand() % (TILE_MAP_SIZE_X);
-	m_pos.y = TILE_MAP_START_POS_Y + rand() % (TILE_MAP_SIZE_Y);
+	m_pos.x = (float)TILE_MAP_START_POS_X + rand() % (TILE_MAP_SIZE_X);
+	m_pos.y = (float)TILE_MAP_START_POS_Y + rand() % (TILE_MAP_SIZE_Y);
 
 	m_bodySize = m_img->GetWidth();
-	m_shape.left = m_pos.x - m_bodySize / 2;
-	m_shape.top = m_pos.y - m_bodySize / 2;
-	m_shape.right = m_pos.x + m_bodySize / 2;
-	m_shape.bottom = m_pos.y + m_bodySize / 2;
+	m_shape.left = (long)m_pos.x - m_bodySize / 2;
+	m_shape.top = (long)m_pos.y - m_bodySize / 2;
+	m_shape.right = (long)m_pos.x + m_bodySize / 2;
+	m_shape.bottom = (long)m_pos.y + m_bodySize / 2;
 
 	mb_isAlive = true;
 	mb_renderScore = false;
@@ -36,9 +36,7 @@ void GranadeItem::Update()
 
 	if (CheckCollision())
 	{
-		GameManager::GetSingleton()->TotKillTankNrCalculation(m_gameScene->GetEnemyMgr()->GetVecEnemys().size());
-		//cout << m_gameScene->GetEnemyMgr()->GetVecEnemys().size() << endl;
-		//cout << GameManager::GetSingleton()->GetKillCount()->totKillTankNr << endl;
+		GameManager::GetSingleton()->TotKillTankNrCalculation((int)m_gameScene->GetEnemyMgr()->GetVecEnemys().size());
 		//mb_isAlive = false;
 		for (int i = 0; i < m_gameScene->GetEnemyMgr()->GetVecEnemys().size(); i++)
 		{
@@ -68,9 +66,9 @@ void GranadeItem::Render(HDC hdc)
 {
 	if (!mb_isAlive) return;
 	if (mb_isShow)
-		m_img->Render(hdc, m_pos.x, m_pos.y);
+		m_img->Render(hdc, (int)m_pos.x, (int)m_pos.y);
 
 	if (mb_renderScore == true)
-		m_img->Render(hdc, m_pos.x, m_pos.y, 4, 0, 2.0f);
+		m_img->Render(hdc, (int)m_pos.x, (int)m_pos.y, 4, 0, 2.0f);
 }
